Integration test code
They are marked as skipped because of a high failure rate with the current way
the tests are written. Perhaps in the future I'll figure out how to make it
work, but having multiple UDP sockets bound and connecting in the same progress
seems to incur some very, very annoying heisenbugs when using Python and QT.

Using `LocalGameServerTest.py` it's fairly straight forward to make them run
reliably against a server that is pre-started.