Simplify smurf-detection logic
This was quite flawed before. It would select the id of rows that have the same uniqueId as the user logging in (including themselves), take the first result, and, if that was a different user, whinge about it.
It would proceed to run a second query to find the username of the second account with which to print an error, and had code for the clearly unreachable case of finding a user id with no associated username but an associated uniqueId that collides with this one... Wat.

By embedding the uid check into the query we prevent false-negatives caused by your own account arbitrarily ending up as the first result.
By selecting the `login` field in the first query, we can eliminate the second query.