Support for friends-only games
Passwords and visibility are now independent. Clients may use a new "visibility" field to request that games be shown only to their friends, independently of having a password.
Passwordedness shall now be checked by the existence of the "password" field, not by the now-removed "access" field.

The server handles the task of not telling clients that aren't on the right friends list about the existence of a friends-only game, however a new "visibility" field is included in
game info objects to allow clients to do fancy things.