Don't send games in the "closed" state to the client
The client discards games in the ENDED or INITIALIZING states.
A recentish server change seems to have greatly increased the
number of such games (or perhaps simply changed how the server
is reporting them so the client is now being told about them).

Due to a stupidity in the client, this causes it to freeze for a
rather long time when it gets a huge pile of junk games dumped
on it on startup. We can save ourself some resources by not doing
so anyway, so let's do that.

We should also make the client not be stupid.

Fixes client/348 from the server direction.