Improve the way the server reports bans and login failures
In order to avoid localisation issues, we should avoid sending
strings to show in the UI over the protocol.

Here, the global exception handler around all message handlers
that listens for login failures is moved into the login function.
Its behaviour is changed to not send a string to the client.
Bans now report deltas, not naive timestamp-ignorant expiry dates.
Unit tests for bans and invalid passwords have been added.