Don't add game connection if failed host connection
After connecting to the host, we don't check if we connected
successfully. This can cause adding the game connection after abort -
since abort is called when replacing a player's game connection and it's
safeguarded not to run twice, this can leave us a dangling game
connection.
Check if we aborted while connecting to host and return if we did.

Signed-off-by: Igor Kotrasinski <ik345186@students.mimuw.edu.pl>