Fix coop scores not getting posted to leaderboard?
Just by looking at the code (someone with database access can verify this for me) it looks like the LIKE query was not matching  column correctly. From #152, filename has the format 'maps/name.version.zip' which is the same format as self.game.map_file_path. So the old query was checking for something like: '%/maps/name.version.zip.%' which clearly doesn't match the entries in the db.