[WIP] Support for multiple player ratings (#486)
* Get rid of BasePlayer class

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Remove Player class mocking in tests

The Player class is essentially plain old data right now. Remove all
mocking of it to make it easier to modify. If we ever bless Player with
non-trivial behaviour, we can make a mock or an abstract class for it.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Improve game stats coverage

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Use a dict-like object for player ratings

Preparations for adding new rating types.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Add asynctest to dependencies

This will save us some manual mucking around with asyncio's sleep.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Replace CoroMock with asynctest's CoroutineMock

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Remove BaseGame class

Everything inherits from the Game class anyway, and all "abstract"
methods are already implemented there.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Fix ban test by formatting >100 year bans as 'forever'

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Make the DB non-global

This should make tests a bit more manageable.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Use asynctest/pytest-asyncio for tests

This replaces homegrown tools for coroutine testing with asynctest and
pytest-asyncio. One benefit is that each test gets its own fresh event
loop, another will come with proper tools for fast-forwarding event
loops.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Add an event loop fast-forwarding decorator

This removes the need for ugly asyncio.sleep monkeypatching. It also
exposes some issues with tests not closing the server after the test
ends; we ignore it for now, eventually we'll fix that too.

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Fix some sleep usage in tests

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* A bit more test coverage

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Remove no longer needed loop fixture

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>

* Update game docstring

Signed-off-by: Igor Kotrasinski <i.kotrasinsk@gmail.com>