Add `connected` field to QDataStreamProtocol to make cleanup easier
This makes it so that cleanup still happens only in one place in 
ServerContext, but other callers of `send_message` will also get an 
error if the protocol has disconnected. So for instance if the ladder 
service tries to send a message start a game but one client disconnects, 
`DisconnectedError` will be raised so that ladder service can 
approprately cancel the game setup, and the ServerContext will clean up 
the connection as soon as the disconnect is detected.