Make all active game gauges and counters a single gauge, counting games managed by game_service
and tag it with game_mode and GameState using only set but not inc or dec.
Whenever the game_service adds or removes a game or whenever a game updates its GameStatus,
all gauges are refreshed by calling GameService.update_active_game_metrics.
To facilitate this, Game.state is now a property that is set via Game.set_state.
Also replaced all '==' GameState enum comparisons with 'is' comparisons.