Issue/#639 Show ban duration to user with higher resolution (#645)
* Show ban duration to user with higher resolution

Until now, a temporarily banned user received ban messages that only
approximated how long they were banned for. For instance, a ban that
lasts another 44 hours would be described as "1 day." Obviously this
would lead to some confusion.

This commit will instead say the exact duration. For instance, "1 day
and 17.5 hours."

Future improvement is to make this rendering happen on the client side
and be translated, but this is a good improvement for the time being.

* Format ban messages as HTML

They were being formatted as strings but it turns out they need to use
HTML linebreaks to be displayed properly.

* Refactor ban errors to be a type of exception