Issue/#721 non ascii in matchmaker game names (#722)
* Move all strategies to their own file

* Rewrite start_game tests to use hypothesis

Note that test_start_game_with_teams will fail.

* Bypass the ascii check when setting matchmaker game names

Since the name was generated internally, we can trust it to be ok. The 
ascii
check is really meant to prevent encoding errors accross different 
system
configurations and to encourage the use of English titles. Clan names 
will
be guaranteed to be valid utf8, and the rest of the game title will be 
in
English since we created it that way.

* Refactor player_factory to generate connection mocks with spec or autospec

* Update codecov to get statuses to post