Fix failing tests
The test was failing because LobbyConnection.on_connection_lost would
take a long time to complete when calling
GameConnection.on_connection_lost. While the connection was waiting for
game results to be processed, a new connection would try to sign in and
join the queue. However, since the PartyService.on_connection_lost logic
hadn't been called yet, it would return the old party for the still
disconnecting player who's state would still be 'Playing'. This might
also be fixed by reworking how Player.__hash__ works, but that's a task
for the future.